Event tracing: put the capability in the block marker, omit it from the events This makes events smaller and tracing quicker, and speeds up reading and sorting the trace file.
HEADS UP: this changes the format of event log files.  Corresponding
changes to the ghc-events package are required (and will be pushed
soon).  Normally we would make backwards-compatible changes, but this
changes the format of every event (to remove the capability) so I'm
breaking the rules this time.  This will be the only time we can do
this, since the format becomes public in 6.12.1.