Fix Trac #3591: very tricky specialiser bug
There was a subtle bug in the interation of specialisation and floating,
described in Note [Specialisation of dictionary functions]. 
  
The net effect was to create a loop where none existed before; plain wrong.
  
In fixing it, I did quite a bit of house-cleaning in the specialiser, and
added a lot more comments.  It's tricky, alas.