Improve error reporting when there's an error inside a spliced expression
If an error occurs in *spliced* code, it can be tricky to understand
what is going on.  With this patch, in
     - epxressions
     - types
we give a helpful indicator that it's in the result of a splice.

For declarations it's harder, because they get type-checked in a 
group with other non-spliced decls, so I have not made the same improvement.
But it's still better than it was.