Break recursive imports between the RTS and base library on Windows.
The file rts/win32/libHSbase.def contains a list of all the 
symbols from the base library that the RTS needs. 

When building the RTS into a DLL on Windows, we want to link the DLL
for the RTS before we link the DLL for the base library. We use 
libHSbase.def to make the "import library" libHSbase.so.a, which 
contains stubs for each of the symbols from the base library that
the RTS needs.